requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-jose[cryptography]>=3.3
pydantic-settings>=2.0
supabase>=2.0
python-multipart>=0.0.9
asyncpg>=0.29
openai>=1.0
